requirements.txt
requests
rich
langchain
langchain-aws
python-dotenv
arize-phoenix-otel
openinference-instrumentation-langchain

